% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_functions.R
\name{fig_configure}
\alias{fig_configure}
\title{Configure the Global Fig Instance}
\usage{
fig_configure(env_prefix, split_on)
}
\arguments{
\item{env_prefix}{(character) A prefix to be prepended to a key before
system environment lookup. Pass an empty string to reset.}

\item{split_on}{(character) A value to split keys on. See Details
section in \code{\link{Fig}}. Providing an empty string disables this behavior.}
}
\value{
Reference to the global fig instance. Other methods can be chained
after this one.
}
\description{
This function allows modifying configuration of the global fig.
Refer to argument descriptions for available options.
}
\details{
Unset arguments do not change configuration.
}
\examples{
fig_configure(env_prefix = "foo_")
fig_configure(split_on = "")
fig_configure() # has no effect
}
