% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_functions.R
\name{fig_delete}
\alias{fig_delete}
\alias{fig_delete_all}
\title{Delete Stored Values}
\usage{
fig_delete(...)

fig_delete_all()
}
\arguments{
\item{...}{Keys to be deleted.}
}
\value{
Reference to the global fig instance. Other methods can be chained
after this one.
}
\description{
These functions allow deleting values stored in the global fig
instance.
}
\examples{
fig_store_many("foo" = 1, "bar" = 2, "baz" = 3)
fig_delete("foo")
fig_delete("bar", "baz")
fig_get_many("foo", "bar", "baz") # == list(NULL, NULL, NULL)
fig_store_many("foo" = 1, "bar" = 2, "baz" = 3)
fig_delete_all()
fig_get_many("foo", "bar", "baz") # == list(NULL, NULL, NULL)
}
