% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIgraph.R
\name{FIgraph}
\alias{FIgraph}
\title{Draw feature interaction graph}
\usage{
FIgraph(
  th = NULL,
  gtype = "S",
  FIobj,
  task = "regression",
  class.name = "all",
  show.edge.weight = TRUE,
  seed = 100
)
}
\arguments{
\item{th}{Threshold for showing edge. If th is NULL, th is mean of interaction values of all pair of teatures}

\item{gtype}{Graph style. 'S' (default) is for static graph and 'I' is for interactive graph.}

\item{FIobj}{Result object from FItable function.}

\item{task}{Type of prediction task, 'regression' (default), 'Classification'}

\item{class.name}{It is for classification task. You can draw feature interaction graph for specific class.}

\item{show.edge.weight}{Decide if edge weight value is displayed or not.}

\item{seed}{According to seed value, The shape of graph is changed.}
}
\value{
Feature interaction graph
}
\description{
This function Draw feature interaction graph.
}
\examples{
library("FIG")
# for regression
data("Boston", package = "MASS")
model <- lm(medv ~ ., data = Boston)
FIobj1 <- FItable(model, train=Boston, target.name="medv", grid=50,
                 task="regression", interaction_type="OH2")
print(FIobj1)
FIgraph(gtype="S", FIobj=FIobj1, task="regression", seed=101)

# for classification
library(e1071)
model2 <- svm(Species~., data=iris)
FIobj2 <- FItable(model2, train=iris, target.name="Species", grid=50,
                 task="classification", interaction_type="OH2", all.class=F)
print(FIobj2)
 FIgraph(th=NULL, gtype="I", FIobj=FIobj2, task="classification", class.name='setosa',
         show.edge.weight=T, seed=102)

# Please ignore warning error when graw graph.

}
