% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FItable.R
\name{FItable}
\alias{FItable}
\title{Create a feature interaction table}
\usage{
FItable(
  model,
  train,
  target.name,
  grid = 50,
  task = "regression",
  interaction_type = "OH2",
  all.class = TRUE
)
}
\arguments{
\item{model}{A prediction model (Classification or regression)}

\item{train}{Training dataset (data frame) that is used to building model}

\item{target.name}{Name of target label name in train dataset}

\item{grid}{Number of grid to calculate partial dependency function. Default is 50)}

\item{task}{Prediction task. "regression" (default) or "classification".}

\item{interaction_type}{Type of feature interaction algorithms. One of "OH2" (default), "FRIEDMAN","GREENWELL".}

\item{all.class}{When task is classification, Feature importance can be calculated for overall classed (all.class==TRUE) or for each class (all.class==FALSE).}
}
\value{
[list] feature interaction & feature importance table
}
\description{
This function create a feature interaction table.
}
\examples{
library("FIG")
# for regression
data("Boston", package = "MASS")
model <- lm(medv ~ ., data = Boston)
FIobj1 <- FItable(model, train=Boston, target.name="medv", grid=50,
                 task="regression", interaction_type="OH2")
print(FIobj1)

# for classification
library(e1071)
model2 <- svm(Species~., data=iris)
FIobj2 <- FItable(model2, train=iris, target.name="Species", grid=50,
                 task="classification", interaction_type="OH2", all.class=F)
print(FIobj2)

}
