% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_fig.R
\name{purge}
\alias{purge}
\title{Purge Values Stored in Global Fig}
\usage{
purge()
}
\description{
This function replaces internal fig object with a new instance
of \code{Fig} class, therefore giving a \verb{fig_*} functions a fresh start.
}
