% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{Fig}
\alias{Fig}
\alias{fig_get}
\alias{fig_set}
\title{Create a Fig to Store Config}
\usage{
fig_get(key)

fig_set(key, value)
}
\arguments{
\item{key}{A key value to store a value for.}

\item{value}{A value to be stored.}
}
\description{
\code{Fig} class is a main driver of this package. For usage details
refer to \code{Fig} methods documentation.

Fig provides a set of exported functions. This makes \code{Fig} class instance
creation optional, and makes the package mimic being a class instance. Those
functions are simple wrappers on an internal fig object and are prefixed to
avoid masking.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Fig$new()}}
\item \href{#method-get}{\code{Fig$get()}}
\item \href{#method-set}{\code{Fig$set()}}
\item \href{#method-clone}{\code{Fig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Fig instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Get a stored value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$get(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A key value to retrieve stored value for.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function returns values based on a following priority
(highest to lowest). If value is not found, then it looks up next level
in the precedence.
\enumerate{
\item System environment variable (case sensitive)
\item Value manually set
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Store a value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$set(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A key value to store a value for.}

\item{\code{value}}{A value to be stored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
