% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{Fig}
\alias{Fig}
\alias{fig_configure}
\alias{fig_delete}
\alias{fig_get}
\alias{fig_store_list}
\alias{fig_delete_all}
\alias{fig_store}
\alias{fig_store_many}
\title{Create a Fig to Store Config}
\usage{
fig_configure(env_prefix)

fig_delete(...)

fig_get(key, split = getOption("fig.split", TRUE))

fig_store_list(cfg, delete_all = FALSE, split = getOption("fig.split", TRUE))

fig_delete_all()

fig_store(key, value, split = getOption("fig.split", TRUE))

fig_store_many(..., .delete_all = FALSE, .split = getOption("fig.split", TRUE))
}
\arguments{
\item{env_prefix}{(character) A prefix to be prepended to a key before system
environment lookup. Pass an empty string to reset.}

\item{...}{Named values. Names are used as keys for provided values.}

\item{key}{A key to store a value for.}

\item{split}{A logical determining whether dots in \code{key} are treated
specially or as is. See Details section.}

\item{cfg}{(Named list) Names are used as keys for provided values.}

\item{delete_all}{A logical determining whether to call \code{delete_all()} before saving
values.}

\item{value}{A value to be stored.}

\item{.delete_all}{A logical determining whether to call \code{delete_all()} before saving
values.}

\item{.split}{A logical determining whether dots in \code{key} are treated
specially or as is. See Details section in \code{store()}.}
}
\description{
\code{Fig} class is a main driver of this package. For usage details
refer to \code{Fig} methods documentation.

Fig provides a set of exported functions. This makes \code{Fig} class instance
creation optional, and makes the package mimic being a class instance. Those
functions are simple wrappers on an internal \code{Fig} object and are prefixed to
avoid masking.
}
\examples{

## ------------------------------------------------
## Method `Fig$new`
## ------------------------------------------------

fig <- Fig$new()
fig <- Fig$new("RCONNECT_")

## ------------------------------------------------
## Method `Fig$configure`
## ------------------------------------------------

fig <- Fig$new()
fig$configure(env_prefix = "RCONNECT_")

## ------------------------------------------------
## Method `Fig$delete`
## ------------------------------------------------

fig <- Fig$new()
fig$store("foo", 1)$store("bar", 2)$store("baz", 3)
fig$get("foo") # == 1
fig$delete("foo")
fig$get("foo") # == NULL
fig$delete("bar", "baz")
fig$get("bar") # == NULL
fig$get("baz") # == NULL

## ------------------------------------------------
## Method `Fig$get`
## ------------------------------------------------

fig <- Fig$new()
fig$store("foo", 1)
fig$get("foo")

fig$store("bar", list(baz = 2))
fig$get("bar.baz")

fig$store("bar.baz", 3)
fig$get("bar.baz") # == 2
fig$get("bar.baz", split = FALSE) # == 3

## ------------------------------------------------
## Method `Fig$store`
## ------------------------------------------------

fig <- Fig$new()
fig$store("foo", 1)
fig$store("bar", 123)$store("baz", list(1, 2, 3))

fig$store("x.y", "a", FALSE)

## ------------------------------------------------
## Method `Fig$store_list`
## ------------------------------------------------

fig <- fig$New()
fig$store_list(list(foo = 1, bar = 2))
fig$store_list(list(foo = 123, baz = "abc"), TRUE)

## ------------------------------------------------
## Method `Fig$store_many`
## ------------------------------------------------

fig <- Fig$new()
fig$store_many("foo" = 1, "bar" = 2)
fig$store_many("foo.bar.baz" = 1, .split = TRUE)
fig$store_many("foo" = "a", "baz" = 123, .delete_all = TRUE, .split = TRUE)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Fig$new()}}
\item \href{#method-configure}{\code{Fig$configure()}}
\item \href{#method-delete}{\code{Fig$delete()}}
\item \href{#method-delete_all}{\code{Fig$delete_all()}}
\item \href{#method-get}{\code{Fig$get()}}
\item \href{#method-store}{\code{Fig$store()}}
\item \href{#method-store_list}{\code{Fig$store_list()}}
\item \href{#method-store_many}{\code{Fig$store_many()}}
\item \href{#method-clone}{\code{Fig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Fig instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$new(env_prefix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env_prefix}}{(character) A prefix to be prepended to a key before
system environment lookup.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fig <- Fig$new()
fig <- Fig$new("RCONNECT_")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configure"></a>}}
\if{latex}{\out{\hypertarget{method-configure}{}}}
\subsection{Method \code{configure()}}{
Change Fig Settings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$configure(env_prefix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env_prefix}}{(character) A prefix to be prepended to a key before system
environment lookup. Pass an empty string to reset.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fig <- Fig$new()
fig$configure(env_prefix = "RCONNECT_")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete stored values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$delete(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Keys to delete values for.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fig <- Fig$new()
fig$store("foo", 1)$store("bar", 2)$store("baz", 3)
fig$get("foo") # == 1
fig$delete("foo")
fig$get("foo") # == NULL
fig$delete("bar", "baz")
fig$get("bar") # == NULL
fig$get("baz") # == NULL
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_all"></a>}}
\if{latex}{\out{\hypertarget{method-delete_all}{}}}
\subsection{Method \code{delete_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$delete_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Get a value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$get(key, split = getOption("fig.split", TRUE))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A key to retrieve its corresponding value.}

\item{\code{split}}{A logical determining whether dots in \code{key} are treated
specially or as is. See Details section.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function returns values based on a following priority
(highest to lowest). If value is not found, then it looks up next level
in the precedence.
\enumerate{
\item System environment variable (case sensitive)
\item Value manually set
}

For system environment lookup dots are replaced by underscores, e.g.
\code{fig$get("foo.bar")} will look up \strong{foo_bar}.

Additionally, Fig treats dots in \code{key} as nest level delimiters.
Therefore, \code{fig$get("foo.bar")} is equivalent to \code{fig$get("foo")$bar}.
This behavior can be disabled either by setting \code{options(fig.split = FALSE)} or by providing \code{split = FALSE} argument.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fig <- Fig$new()
fig$store("foo", 1)
fig$get("foo")

fig$store("bar", list(baz = 2))
fig$get("bar.baz")

fig$store("bar.baz", 3)
fig$get("bar.baz") # == 2
fig$get("bar.baz", split = FALSE) # == 3
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store"></a>}}
\if{latex}{\out{\hypertarget{method-store}{}}}
\subsection{Method \code{store()}}{
Store a value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$store(key, value, split = getOption("fig.split", TRUE))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A key to store a value for.}

\item{\code{value}}{A value to be stored.}

\item{\code{split}}{A logical determining whether dots in \code{key} are treated
specially or as is. See Details section.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fig treats dots in \code{key} as nest level delimiters. Therefore,
\code{fig$store("foo.bar", 1)} is equivalent to \verb{fig$store("foo", list(bar = 1)}.
This behavior can be disabled either by setting \code{options(fig.split = FALSE)} or by providing \code{split = FALSE} argument.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fig <- Fig$new()
fig$store("foo", 1)
fig$store("bar", 123)$store("baz", list(1, 2, 3))

fig$store("x.y", "a", FALSE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store_list"></a>}}
\if{latex}{\out{\hypertarget{method-store_list}{}}}
\subsection{Method \code{store_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$store_list(cfg, delete_all = FALSE, split = getOption("fig.split", TRUE))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfg}}{(Named list) Names are used as keys for provided values.}

\item{\code{delete_all}}{A logical determining whether to call \code{delete_all()} before
saving values.}

\item{\code{split}}{A logical determining whether dots in \code{key} are treated
specially or as is. See Details section in \code{store()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fig <- fig$New()
fig$store_list(list(foo = 1, bar = 2))
fig$store_list(list(foo = 123, baz = "abc"), TRUE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store_many"></a>}}
\if{latex}{\out{\hypertarget{method-store_many}{}}}
\subsection{Method \code{store_many()}}{
Set any number of values at once
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$store_many(..., .delete_all = FALSE, .split = getOption("fig.split", TRUE))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named values. Names are used as keys for provided values.}

\item{\code{.delete_all}}{A logical determining whether to call \code{delete_all()} before
saving values.}

\item{\code{.split}}{A logical determining whether dots in \code{key} are treated
specially or as is. See Details section in \code{store()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fig <- Fig$new()
fig$store_many("foo" = 1, "bar" = 2)
fig$store_many("foo.bar.baz" = 1, .split = TRUE)
fig$store_many("foo" = "a", "baz" = 123, .delete_all = TRUE, .split = TRUE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
